+++ full-implementation/src/components/Guns.cpp
