+++ full-implementation/src/components/Guns.h
