+++ full-implementation/src/components/Player.cpp
@@ -1,7 +1,4 @@
 #include "Player.h"
-#include "Constants.h"
-#include "Map.h"
-#include "Vector.h"
 #include <SDL2/SDL.h>
 #include <SDL2/SDL_image.h>
 #include <SDL2/SDL_render.h>
