+++ full-implementation/src/components/Sprites.cpp
