+++ full-implementation/src/maps/Map.cpp
@@ -1,7 +1,5 @@
 #include "Map.h"
-#include "Constants.h"
-#include "MapParser.h"
-#include "ObjectParser.h"
+#include "../utils/Constants.h"
 #include <SDL2/SDL.h>
 #include <SDL2/SDL_image.h>
 #include <algorithm>
