+++ full-implementation/src/maps/ObjectParser.cpp
